services/notification_service/requirements.txt
### FILE: services/notification_service/requirements.txt
grpcio
grpcio-tools
protobuf
pika
